+++ pytc/global_connector/__init__.py
@@ -6,5 +6,8 @@ __author__ = "Michael J. Harms"
 __date__ = "2016-12-19"
 __all__ = [] 
 
-from .connector import GlobalConnector, NumProtons, VantHoff, VantHoffExtended
-from .base import GlobalFit
+from .base import GlobalConnector
+from .num_protons import NumProtons
+from .vant_hoff import VantHoff
+from .vant_hoff_extended import VantHoffExtended
+
